Refactor AppInputPriceControl with AppInputNumberControl.
